Add a lint check for innerHTML
